jobber_zapier_app/requirements.txt
Flask
requests
gspread
oauth2client
python-dotenv
gunicorn
